# docs/swagger.yaml
basePath: /
definitions:
...
  serializer.TireJSON:
    properties:
      description:
        type: string
      is_delete:
        type: boolean
      photo:
        type: string
      tire_id:
        type: integer
      tire_material_coefficient:
        type: number
      tire_thickness_coefficient:
        type: number
      tire_title:
        type: string
      video:
        type: string
    type: object
...
  serializer.TirePressureEntryUpdateJSON:
    properties:
      coating_coefficient:
        type: number
      direction:
        type: string
      pressure:
        type: number
    type: object
...
  serializer.TirePressureListJSON:
    properties:
      air_temperature:
        type: number
      car_weight:
        type: number
      creator_login:
        type: string
      date_completed:
        type: string
      date_create:
        type: string
      date_formed:
        type: string
      moderator_login:
        type: string
      status:
        type: string
      tire_entries_count:
        type: integer
      tire_pressure_id:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API для расчёта давления в шинах
  title: Tire Pressure API
  version: "1.0"
paths:
  /api/tire-pressure-entries/{tire_id}/{tire_pressure_id}:
    delete:
      parameters:
      - description: ID шины
        in: path
        name: tire_id
        required: true
        type: integer
      - description: ID заявки
        in: path
        name: tire_pressure_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Обновлённая заявка
          schema:
            $ref: '#/definitions/serializer.TirePressureListJSON'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Удалить шину из заявки
      tags:
      - tire-pressure-entries
    put:
      consumes:
      - application/json
      description: Коэффициент покрытия, давление и другие параметры для пары шина–заявка.
      parameters:
      - description: ID шины
        in: path
        name: tire_id
        required: true
        type: integer
      - description: ID заявки
        in: path
        name: tire_pressure_id
        required: true
        type: integer
      - description: Поля для обновления
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/serializer.TirePressureEntryUpdateJSON'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/serializer.TirePressureEntryUpdateJSON'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Обновить запись в заявке
      tags:
      - tire-pressure-entries
  /api/tire-pressure-entries/add/{tire_id}:
    post:
      description: Добавляет шину в черновик заявки; при первом добавлении создаётся
        заявка (201 + Location). Поддерживает JSON и form-data.
      parameters:
      - description: ID шины
        in: path
        name: tire_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Успешно добавлено
          schema:
            additionalProperties: true
            type: object
        "201":
          description: Заявка создана
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Уже в заявке
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Добавить шину в заявку
      tags:
      - tire-pressure-entries
...
  /api/tires:
    get:
      parameters:
      - description: Поиск по названию шины
        in: query
        name: query
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/serializer.TireJSON'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Список шин
      tags:
      - tires
    post:
      consumes:
      - multipart/form-data
      description: multipart/form-data или application/json; только для модераторов
      parameters:
      - description: Название шины
        in: formData
        name: tire_title
        type: string
      - description: Коэффициент материала
        in: formData
        name: tire_material_coefficient
        type: number
      - description: Коэффициент толщины
        in: formData
        name: tire_thickness_coefficient
        type: number
      - description: Описание
        in: formData
        name: description
        type: string
      - description: Фото шины
        in: formData
        name: photo
        type: file
      - description: Видео шины
        in: formData
        name: video
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/serializer.TireJSON'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Требуется роль модератора
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Создать шину
      tags:
      - tires
...
  /tires:
    get:
      parameters:
      - description: Поиск по названию шины
        in: query
        name: query
        type: string
      produces:
      - text/html
      responses:
        "200":
          description: HTML-страница
          schema:
            type: string
      summary: Главная страница (список шин)
      tags:
      - tires-html
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"